+++ packages/components/atoms/f-card/test/component/f-card.component.spec.js
@@ -1,14 +1,11 @@
-const Card = require('../../test-utils/component-objects/f-card.component');
-
-const card = new Card();
+import Card from '../../test-utils/component-objects/f-card.component';
 
 describe('f-card component tests', () => {
-    beforeEach(async () => {
-        await card.load();
-    });
-
     it('should display the f-card component', async () => {
+        // Arrange
+        await Card.load();
+        
         // Assert
-        await expect(await card.isComponentDisplayed()).toBe(true);
+        await expect(await Card.isComponentDisplayed()).toBe(true);
     });
 });