+++ packages/components/atoms/f-form-field/test-utils/component-objects/f-form-field--visual.component.js
@@ -1,9 +1,11 @@
-const Page = require('@justeat/f-wdio-utils/src/base.page');
+import Page from '@justeat/f-wdio-utils/src/base.page';
 
-module.exports = class FormField extends Page {
+class FormField extends Page {
     constructor () {
         super('atom-folder', 'f-form-field--form-field-component');
     }
 
     get component () { return $('[data-test-id="formfield-container"]'); }
 };
+
+export default new FormField();