+++ packages/components/atoms/f-image-tile/test-utils/component-objects/f-image-tile.component.js
@@ -1,6 +1,6 @@
-const Page = require('@justeat/f-wdio-utils/src/base.page');
+import Page from'@justeat/f-wdio-utils';
 
-module.exports = class ImageTile extends Page {
+class ImageTile extends Page {
     constructor () {
         super('atom', 'image-tile-component');
     }
@@ -9,3 +9,5 @@ module.exports = class ImageTile extends Page {
         return this.component.isClickable();
     }
 };
+
+export default new ImageTile();