+++ packages/components/atoms/f-popover/test-utils/component-objects/f-popover.component.js
@@ -1,7 +1,9 @@
-const Page = require('@justeat/f-wdio-utils/src/base.page');
+import Page from '@justeat/f-wdio-utils';
 
-module.exports = class Popover extends Page {
+class Popover extends Page {
     constructor () {
         super('atom', 'popover-component');
     }
 };
+
+export default new Popover();