+++ packages/components/atoms/f-spinner/test-utils/component-objects/f-spinner.component.js
@@ -1,7 +1,8 @@
-const Page = require('@justeat/f-wdio-utils/src/base.page');
-
-module.exports = class Spinner extends Page {
+import Page from '@justeat/f-wdio-utils';
+class Spinner extends Page {
     constructor () {
         super('atom', 'v-spinner-component');
     }
 };
+
+export default new Spinner();